kube-spawn: split out commands into multiple files
For better maintainance, we should split out kube-spawn.go into
multiple files, so that each file can take care of its corresponding
command, e.g. init.go for init, setup.go for setup.
