+++ cmd/kube-spawn/up.go
@@ -1,0 +1,41 @@
+/*
+Copyright 2017 Kinvolk GmbH
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+var (
+	cmdUp = &cobra.Command{
+		Use:   "up",
+		Short: "Set up nodes bringing up nspawn containers (deprecated)",
+		Run:   runUp,
+	}
+)
+
+func init() {
+	cmdKubeSpawn.AddCommand(cmdUp)
+}
+
+func runUp(cmd *cobra.Command, args []string) {
+	log.Printf("WARNING: up command is deprecated. Please run setup command instead.")
+	os.Exit(1)
+}
